refactor: move migrate_user endpoints to dedicated api file
