+++ propelauth_py/api.py
@@ -6,52 +6,9 @@ import requests
 from requests.auth import AuthBase
 
 from propelauth_py.errors import (
-    BadRequestException,
-    UserNotFoundException,
     EndUserApiKeyException,
     EndUserApiKeyNotFoundException,
 )
-
-
-def _migrate_user_from_external_source(
-    auth_url,
-    integration_api_key,
-    email,
-    email_confirmed,
-    existing_user_id=None,
-    existing_password_hash=None,
-    existing_mfa_base32_encoded_secret=None,
-    ask_user_to_update_password_on_login=False,
-    enabled=None,
-    first_name=None,
-    last_name=None,
-    username=None,
-    properties=None,
-):
-    url = auth_url + "/api/backend/v1/migrate_user/"
-    json = {
-        "email": email,
-        "email_confirmed": email_confirmed,
-        "existing_user_id": existing_user_id,
-        "existing_password_hash": existing_password_hash,
-        "existing_mfa_base32_encoded_secret": existing_mfa_base32_encoded_secret,
-        "ask_user_to_update_password_on_login": ask_user_to_update_password_on_login,
-        "enabled": enabled,
-        "first_name": first_name,
-        "last_name": last_name,
-        "username": username,
-        "properties": properties,
-    }
-
-    response = requests.post(url, json=json, auth=_ApiKeyAuth(integration_api_key))
-    if response.status_code == 401:
-        raise ValueError("integration_api_key is incorrect")
-    elif response.status_code == 400:
-        raise BadRequestException(response.json())
-    elif not response.ok:
-        raise RuntimeError("Unknown error when migrating user")
-
-    return response.json()
 
 
 def _fetch_api_key(auth_url, integration_api_key, api_key_id):
